(PC-26134) refactor(search): export DEFAULT_RADIUS constant to constants file
