+++ src/features/search/constants.ts
@@ -4,3 +4,4 @@ export const MIN_HISTORY_RESULTS = 2
 export const LIST_ITEM_HEIGHT = 130
 export const SCROLL_TO_TOP_BUTTON_LINEAR_GRADIENT_START = { x: 0, y: 0 }
 export const SCROLL_TO_TOP_BUTTON_LINEAR_GRADIENT_END = { x: 0, y: 1 }
+export const DEFAULT_RADIUS = 50
